add doc and mobile component

+++ src/com/wisedu/emap/itpub/util/GenTemplateUtil.java
@@ -1,10 +1,15 @@
 package com.wisedu.emap.itpub.util;
 
+import java.io.BufferedReader;
+import java.io.BufferedWriter;
 import java.io.File;
+import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.HashMap;
+import java.util.Iterator;
 import java.util.Map;
+import java.util.Map.Entry;
 
 import freemarker.template.Configuration;
 import freemarker.template.TemplateExceptionHandler;
@@ -71,7 +76,7 @@ public class GenTemplateUtil {
 
 		String appName = (String) params.get("appName");
 
-		// 生成classpath文件,继承itpu、引用emapflow
+		// 生成classpath文件,继承itpub、引用emapflow
 		File classpath = new File(PROJECT_PATH + appName + "/" + ".classpath");
 		if (!classpath.getParentFile().exists()) {
 			return "无法找到应用：" + appName;
@@ -81,7 +86,6 @@ public class GenTemplateUtil {
 		File controllerFile = new File(PROJECT_PATH + appName + "/src/com/wisedu/emap/" + appName + "/controller/"
 				+ "AppIndexController.java");
 
-		// 删除原来的index目录
 		// 重写config.js和index.jsp
 		File configjs = new File(PROJECT_PATH + appName + "/web/" + "config.js");
 		File indexjsp = new File(PROJECT_PATH + appName + "/web/" + "index.jsp");
@@ -95,7 +99,21 @@ public class GenTemplateUtil {
 		if (!commonjsp.getParentFile().exists()) {
 			commonjsp.getParentFile().mkdirs();
 		}
+		// 判断是否已存在，若存在，则不再生成
+		Map<String, Boolean> isExist = new HashMap<String, Boolean>();
+		isExist.put("AppIndexController.java", controllerFile.exists());
+		isExist.put("index.jsp", indexjsp.exists());
+		isExist.put("common.jsp", commonjsp.exists());
+		isExist.put("source.jsp", sourcejsp.exists());
 
+		Iterator<Entry<String, Boolean>> ie = isExist.entrySet().iterator();
+		while (ie.hasNext()) {
+			Map.Entry<String, Boolean> entry = (Map.Entry<String, Boolean>) ie.next();
+			if (entry.getValue() == true) {
+				return "模板文件" + entry.getKey() + "已存在，若需重新生成，请先删除模板文件！";
+			}
+		}
+		// 生成的操作
 		cfg.getTemplate("classpath.ftl").process(params, new FileWriter(classpath));
 		cfg.getTemplate("controller.ftl").process(params, new FileWriter(controllerFile));
 		cfg.getTemplate("configjs.ftl").process(params, new FileWriter(configjs));
@@ -103,7 +121,56 @@ public class GenTemplateUtil {
 		cfg.getTemplate("commonjsp.ftl").process(params, new FileWriter(commonjsp));
 		cfg.getTemplate("sourcejsp.ftl").process(params, new FileWriter(sourcejsp));
 
-		return "生成itpub模板文件成功！";
+		// 在EMAP_APP末尾添加EMAP应用继承关系
+		File emap_app = new File(PROJECT_PATH + appName + "/EMAP_APP");
+		if (!emap_app.exists())
+			return "EMAP_APP文件不存在，请检查应用目录！";
+		else
+			addEmapParent(emap_app);
+
+		return "itpub模板文件生成成功！";
+
+	}
+
+	private void addEmapParent(File emap_app) {
+		FileReader fr = null;
+		BufferedReader br = null;
+		FileWriter fw = null;
+		BufferedWriter bw = null;
+		try {
+			fr = new FileReader(emap_app);
+			br = new BufferedReader(fr);
+			String content = "";
+			while (br.ready()) {
+				content += br.readLine() + "\n";
+			}
+			content += "parent=itpub";
+
+			fw = new FileWriter(emap_app);
+			bw = new BufferedWriter(fw);
+			bw.write(content);
+			bw.flush();
+
+		} catch (Exception e) {
+			e.printStackTrace();
+		} finally {
+			try {
+				if (fr != null) {
+					fr.close();
+				}
+				if (br != null) {
+					br.close();
+				}
+				if (fw != null) {
+					fw.close();
+				}
+				if (bw != null) {
+					bw.close();
+				}
+			} catch (Exception e) {
+				e.printStackTrace();
+			}
+		}
 
 	}
 
@@ -125,6 +192,9 @@ public class GenTemplateUtil {
 		if (!daofile.getParentFile().exists()) {
 			daofile.getParentFile().mkdirs();
 		}
+		if (daofile.exists())
+			return "模板文件" + daoName + ".java已存在，若需重新生成，请先删除模板文件！";
+
 		cfg.getTemplate("dao.ftl").process(params, new FileWriter(daofile));
 
 		File implfile = new File(
@@ -132,6 +202,9 @@ public class GenTemplateUtil {
 		if (!implfile.getParentFile().exists()) {
 			implfile.getParentFile().mkdirs();
 		}
+		if (implfile.exists())
+			return "模板文件" + daoName + "_IMPL.java已存在，若需重新生成，请先删除模板文件！";
+
 		cfg.getTemplate("daoimpl.ftl").process(params, new FileWriter(implfile));
 
 		return "生成dao模板文件成功！";
